providing the posthog creation and analytics config via hilt in order to make the analytics impl testable
